refactor(test): extract compile-fail helper and containsComponent matcher to test module
Move assertDoesNotCompile into modules/test as a shared public utility and
containsComponent into ComponentMatchers as shouldContainComponent, eliminating
private duplicates in the minimessage and core test files and consolidating
the kctfork-core dependency in the one place that owns it.
