+++ modules/minimessage/build.gradle
@@ -7,5 +7,4 @@ dependencies {
 
     testImplementation project(':core')
     testImplementation project(':test')
-    testImplementation kotventureDependencies."kctfork-core"
 }
